+++ src/pages/Club/Application/hooks/useGetClubFee.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { getClubFee } from '@/apis/club';
-import { clubQueryKeys } from '@/pages/Club/ClubList/hooks/useGetClubs';
+import { clubQueryKeys } from '@/apis/club/queries';
 
 export const useGetClubFee = (clubId: number) => {
   return useSuspenseQuery({
