+++ src/pages/Club/ClubDetail/hooks/useGetClubDetail.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { getClubDetail } from '@/apis/club';
-import { clubQueryKeys } from '@/pages/Club/ClubList/hooks/useGetClubs';
+import { clubQueryKeys } from '@/apis/club/queries';
 
 export const useGetClubDetail = (clubId: number) => {
   return useSuspenseQuery({
