+++ src/pages/Club/ClubDetail/hooks/useGetClubRecruitment.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { getClubRecruitment } from '@/apis/club';
-import { clubQueryKeys } from '@/pages/Club/ClubList/hooks/useGetClubs';
+import { clubQueryKeys } from '@/apis/club/queries';
 
 const useGetClubRecruitment = (clubId: number) => {
   return useSuspenseQuery({
