+++ src/pages/Club/ClubList/hooks/useGetClubs.ts
@@ -1,24 +1,7 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { getClubs } from '@/apis/club';
 import type { ClubResponse } from '@/apis/club/entity';
-
-export const clubQueryKeys = {
-  all: ['clubs'],
-  list: (params: { limit: number; query?: string; isRecruiting: boolean }) => [
-    ...clubQueryKeys.all,
-    'list',
-    params.limit,
-    params.query,
-    params.isRecruiting,
-  ],
-  detail: (clubId: number) => [...clubQueryKeys.all, 'detail', clubId],
-  members: (clubId: number) => [...clubQueryKeys.all, 'members', clubId],
-  recruitment: (clubId: number) => [...clubQueryKeys.all, 'recruitment', clubId],
-  fee: (clubId: number) => [...clubQueryKeys.all, 'fee', clubId],
-  questions: (clubId: number) => [...clubQueryKeys.all, 'questions', clubId],
-  joined: () => [...clubQueryKeys.all, 'joined'],
-  applied: () => [...clubQueryKeys.all, 'applied'],
-};
+import { clubQueryKeys } from '@/apis/club/queries';
 
 interface UseGetClubsParams {
   limit?: number;
@@ -29,7 +12,7 @@ interface UseGetClubsParams {
 
 export const useGetClubs = ({ limit = 10, query, enabled = true, isRecruiting = false }: UseGetClubsParams = {}) => {
   return useInfiniteQuery({
-    queryKey: clubQueryKeys.list({ limit, query, isRecruiting }),
+    queryKey: clubQueryKeys.infinite.list({ limit, query, isRecruiting }),
     queryFn: ({ pageParam = 1 }) =>
       getClubs({
         page: pageParam,
